blackhaven_framework/requirements.txt
colorama>=0.4.6
requests>=2.31.0
dnspython>=2.6.1
python-whois>=0.8.0
PyYAML>=6.0.1
